Move templates package to pkg/template
